--delete account flow is complete -- need to spruce it up with styling

+++ components/Parent/DeleteAccount.js
@@ -1,5 +1,6 @@
-import { useMutation } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
+import { useMutation } from "@apollo/react-hooks";
+import gql from "graphql-tag";
+import Router from "next/router";
 
 const DELETE_PARENT_ACCOUNT = gql`
   mutation DELETE_PARENT_ACCOUNT {
@@ -9,23 +10,25 @@ const DELETE_PARENT_ACCOUNT = gql`
   }
 `;
 export default function DeleteAccount() {
-  const [deleteParentAccount, { loading, error }] = useMutation(
+  const [deleteParentAccount, { loading, error, client }] = useMutation(
     DELETE_PARENT_ACCOUNT,
     {
-      onCompleted: () => {
-        console.log('account delete mutation complete');
-        // Router.push('/');
-        // client.clearStore();
+      onCompleted: (data) => {
+        console.log("complete : data:", data);
+        client.clearStore();
+        Router.push("/");
       },
     }
   );
   async function deleteAccount() {
-    await deleteParentAccount();
+    if (confirm("Do you really want to delete your account?")) {
+      await deleteParentAccount();
+    }
   }
   return (
     <button
-      type='button'
-      className='btn-danger'
+      type="button"
+      className="btn-danger"
       onClick={() => deleteAccount()}
     >
       Delete My Account
